Rename urlresult* -> general_result* to match class name.
