+++ searchsummary/src/vespa/searchsummary/docsummary/general_result.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "urlresult.h"
+#include "general_result.h"
 #include "resultconfig.h"
 #include <zlib.h>
 #include <cassert>
