+++ searchsummary/src/vespa/searchsummary/docsummary/general_result.h
